Move host functions to new package
